Remove unnecessary abstraction for REST controllers
